Correlation Matrix example is added.
Worldbank population dataset has been transformed. Contry Codes converted to columns and the years indexing the rows.

The correlation matrix of the transformed dataset is visualized with both matplotlib and seaborn. Visual representation shows that the population changed similarly in some countries.

Finally, we defined a threshold for similarity and queried the countries which are similar to Hungary based on the dataset.