seeds yielded clusters are compared with intersection
The cluster labels are assigned randomly. Therefore the different clustering algorithm may give different label for the same record. So the intersection of the clusters of the different clustering methods are determined.

Bug fix: DBSCAN Visualization hid the outlier points. Problem was caused by the range function. The range function was change to the set function in the for loop.